add option to discard the video ratio of the primary video
